+++ projects/mini-rx-store/src/lib/store-core.ts
@@ -100,10 +100,9 @@ class StoreCore {
             initialState?: StateType;
         } = {}
     ) {
-        reducer =
-            config.metaReducers && config.metaReducers.length
-                ? combineMetaReducers<StateType>(config.metaReducers)(reducer)
-                : reducer;
+        reducer = config.metaReducers?.length
+            ? combineMetaReducers<StateType>(config.metaReducers)(reducer)
+            : reducer;
 
         checkFeatureExists(featureKey, this.reducers);
 
@@ -147,11 +146,11 @@ class StoreCore {
             );
         }
 
-        if (config.metaReducers && config.metaReducers.length) {
+        if (config.metaReducers?.length) {
             this.addMetaReducers(...config.metaReducers);
         }
 
-        if (config.extensions && config.extensions.length) {
+        if (config.extensions?.length) {
             const sortedExtensions: StoreExtension[] = sortExtensions(config.extensions);
             sortedExtensions.forEach((extension) => this.addExtension(extension));
         }